# public/openapi-schema.json
{"openapi":"3.0.0","info":{"title":"Microservice Template","description":"Use it to start making Microservice","version":"0.2.1","contact":{},"license":{"name":"MIT","url":""}},"tags":[{"name":"v1","description":""},{"name":"monitoring","description":""}],"servers":[{"url":"http://0.0.0.0:8888"},{"url":"http://localhost:8888"}],"components":{"schemas":{"SampleRequestDto":{"type":"object","properties":{"ping":{"type":"string","description":"Sample string","example":"Foo"}},"required":["ping"]},"SampleResponseDto":{"type":"object","properties":{"pong":{"type":"string","description":"Sample string","example":"Bar"}},"required":["pong"]},"ErrorResponseDto":{"type":"object","properties":{"type":{"type":"string","description":"The type of the Exception","example":"ExampleErrorType"},"message":{"type":"string","description":"The message of the Exception","example":"Details if any"}},"required":["type","message"]},"SampleEntity":{"type":"object","properties":{"sample":{"type":"string","description":"sample"}},"required":["sample"]}}},"paths":{"/health":{"get":{"operationId":"check","summary":"","parameters":[],"responses":{"200":{"description":"The Health Check is successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"ok"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}}},"tags":["monitoring"]}},"/":{"get":{"operationId":"MainController_getStatus","parameters":[],"responses":{"200":{"description":"All right"}},"tags":["v1"]}},"/v1/process":{"post":{"operationId":"SampleProcessorController_process","summary":"Short description","description":"Long description / instructions","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SampleRequestDto"}}}},"responses":{"200":{"description":"All right, no new resources was created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SampleResponseDto"}}}},"201":{"description":"All right, new resources was created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SampleResponseDto"}}}},"default":{"description":"Not all right. HTTP Code will reflect an error type","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["v1"]}},"/v1/bar/{foo}":{"get":{"operationId":"SampleProcessorController_getBarByFoo","summary":"Another short description","parameters":[{"name":"foo","required":true,"in":"path","example":"Foo","schema":{"type":"string"}}],"responses":{"200":{"description":"All right","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SampleEntity"}}}},"204":{"description":"All right, but empty, no message-body"},"default":{"description":"Not all right. HTTP Code will reflect an error type","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponseDto"}}}}},"tags":["v1"]}}}}